binary optimization
remove inlining to reduce size by 22kb.
